-Refactored file structure -Implemented first test
