+++ backend/src/datasources.ts
@@ -1,0 +1,8 @@
+import PostDatasource from "./datasources/post.datasource";
+import UserDatasource from "./datasources/user.datasource";
+
+const dataSources = {
+    postDatasource: new PostDatasource(),
+    userDatasource: new UserDatasource()
+}
+export default dataSources;
