+++ backend/tsconfig.json
@@ -6,7 +6,8 @@
     "outDir": "dist",
     "sourceMap": true,
     "noImplicitAny": false,
-    "skipLibCheck": true
+    "skipLibCheck": true, 
+    "types": ["node", "jest"] 
   },
   "exclude": ["node_modules", "dist"],
   "include": ["src/**/*"]
